Refactor with test utilities + verified it works
