+++ test/test_leap_connection_topics.py
@@ -5,27 +5,9 @@ import launch_ros
 import rclpy
 from launch import LaunchDescription
 from rclpy.node import Node
+from test_utilities.parameter_utility import TopicChecker
 
-class TopicChecker(Node):
-    def __init__(self):
-        super().__init__("topic_checker")
-
-    def topic_exists(self, topic_name):
-        topic_list = self.get_topic_names_and_types()
-        print(topic_list)
-        for topic, _ in topic_list:
-            if topic == topic_name:
-                return True
-        return False
-
-@pytest.fixture(autouse=True, scope="session")
-def initialize_rclpy():
-    # Set an arbitrary ROS_DOMAIN_ID so that the test is performed without inteference
-    os.environ['ROS_DOMAIN_ID'] = '42'
-
-    rclpy.init()
-    yield
-    rclpy.shutdown()
+node_name = "test_leap_node"
 
 @launch_pytest.fixture
 def launch_leap_ros2_node(config_params):
@@ -38,7 +20,7 @@ def launch_leap_ros2_node(config_params):
             launch_ros.actions.Node(
                 package="leap_ros2",
                 executable="leaphand_node.py",
-                name="test_leap_node",
+                name=node_name,
                 output="screen",
                 parameters=[config_params],
             ),
@@ -60,6 +42,4 @@ def test_topic_exists(config_params, topic_checker, topic_name_str):
     """Checks if a topic is being published or subscribed to"""
     topic_to_check = config_params[topic_name_str]
     rclpy.spin_once(topic_checker, timeout_sec=3.0)
-    assert topic_checker.topic_exists(topic_to_check), (
-        f'Topic "{topic_to_check}" not detected'
-    )
+    topic_checker.topic_exists(topic_to_check)
